build(msvc): clean /W4 /WX build (warnings + .lib collision + POSIX shims)
Fixes 62 warnings (C4244/C4100/C4310/C4312/C4701) and 2 build blockers
that prevented the canonical MSVC build:

* C4701 in mqb.c (x13) and btl.c: zero-init locals at declaration.
* C4244 in bhd5.c: add UINT32_MAX overflow check mirroring the DS3 branch
  before truncating unpadded_size to u32.
* C4244 in flver2_decode.c: explicit (uint8_t) on max_uv_count /
  max_color_count assignments.
* C4100 in 4 sources: (void)param; for params kept for API symmetry.
* C4310 in 5 tests: Unity's TEST_ASSERT_EQUAL_HEX{8,16} internally cast
  to signed int{8,16}_t, which truncates 0xFF/0x8000/0xFFFF; switch
  the affected calls to _UINT{8,16}.
* C4312 in test_sf_util.c: cast 0xdeadbeef sentinel through uintptr_t.
* .lib name collision: tag static archive as souls_formats_static.lib
  on MSVC only (MinGW already disambiguates via .a vs .dll.a).
* F_OK / access(): add a _MSC_VER-only POSIX shim in the two
  paramdef e2e tests.

MinGW build and artifact names unchanged. Both toolchains now compile
clean and run 182/182 tests green.

+++ src/geom/flver2_decode.c
@@ -69,10 +69,10 @@ SF_API sf_result_t sf_flver2_decode_mesh(const sf_flver2_t *f, size_t mesh_index
                 case SF_FLVER_LAYOUT_SEMANTIC_BONE_INDICES: has_bone_indices = true; break;
                 case SF_FLVER_LAYOUT_SEMANTIC_BONE_WEIGHTS: has_bone_weights = true; break;
                 case SF_FLVER_LAYOUT_SEMANTIC_UV:
-                    if (member->index + 1 > max_uv_count) max_uv_count = member->index + 1;
+                    if (member->index + 1 > max_uv_count) max_uv_count = (uint8_t)(member->index + 1);
                     break;
                 case SF_FLVER_LAYOUT_SEMANTIC_VERTEX_COLOR:
-                    if (member->index + 1 > max_color_count) max_color_count = member->index + 1;
+                    if (member->index + 1 > max_color_count) max_color_count = (uint8_t)(member->index + 1);
                     break;
                 default: break;
             }
